identified the pb
The pb is that weight is of type double and this does not work for some reason. When we read a graph via text file, weight is transformed into an int and everything works perfectly. So have to keep working on that.
Also, I corrected the main() because we were confusing the number of vertices to create a graph according to the type of graph (random, fixed or via txt)